refactor(provider): перенести provider HTTP API в provider/api
